x86: don't use lddqu
lddqu is a false optimization.

It is only faster in some scenarios on old x86 chips which don't even
support SSE4.1.

It is otherwise equivalent or _slower_ on newer chips.

Additionally, and more importantly, unlike _mm_loadu_si128 which appears
to the compiler as an unaligned memory access, _mm_lddqu_si128 appears
as an opaque intrinsic, and therefore cannot be optimized.